+++ cocos2d/cocos/base/base64.cpp
@@ -40,8 +40,8 @@ int _base64Decode(const unsigned char *input, unsigned int input_len, unsigned c
     unsigned int output_idx = 0;
 
     for (char i = (sizeof alphabet) - 1; i >= 0 ; i--) {
-        inalphabet[alphabet[i]] = 1;
-        decoder[alphabet[i]] = i;
+        inalphabet[alphabet[static_cast<int>(i)]] = 1;
+        decoder[alphabet[static_cast<int>(i)]] = i;
     }
 
     char_count = 0;
