+++ cocos2d/cocos/renderer/CCPass.cpp
@@ -216,7 +216,7 @@ void Pass::updateMVPUniform(const Mat4& modelView)
     if (_locNormalMatrix)
     {
         auto normalMatrix = ccutils::getNormalMat3OfMat4(modelView);
-        _programState->setUniform(_locNormalMatrix, normalMatrix.data(), sizeof(normalMatrix[0]) * normalMatrix.size());
+        _programState->setUniform(_locNormalMatrix, normalMatrix.data(), static_cast<unsigned int>(sizeof(normalMatrix[0]) * normalMatrix.size()));
     }
 
 }
